Fix compilation errors

+++ Green/Main.cpp
@@ -15,7 +15,7 @@ int main(int argc, char** argv)
 	}
 
 	file.seekg(0, file.end);
-	int length = file.tellg();
+	int length = static_cast<int>(file.tellg());
 	file.seekg(0, file.beg);
 
 	unsigned char* buffer = new unsigned char[length];
